+++ R/dispatcher.R
@@ -20,8 +20,8 @@
 #'   dial in to), including the port to connect to e.g. tcp://hostname:5555' or
 #'   'tcp://10.75.32.70:5555'. Specify 'tls+tcp://' to use secure TLS
 #'   connections.
-#' @param n \[default 0L\] if specified, the integer number of daemons to be
-#'   launched locally by the host process.
+#' @param n if specified, the integer number of daemons to be launched locally
+#'   by the host process.
 #'
 #' @return Invisible NULL.
 #'
@@ -62,14 +62,13 @@ dispatcher <- function(host, url = NULL, n = 0L, ...) {
   res <- res[[5L]]
 
   inq <- outq <- list()
-  events <- integer()
   connections <- count <- 0L
   envir <- new.env(hash = FALSE, parent = emptyenv())
   `[[<-`(envir, "stream", res)
 
   if (n) {
     for (i in seq_len(n))
-      while(!until(cv, .limit_long))
+      while (!until(cv, .limit_long))
         cv_signal(cv) || wait(cv) || return()
 
     changes <- read_monitor(m)
@@ -88,7 +87,7 @@ dispatcher <- function(host, url = NULL, n = 0L, ...) {
       url <- sub_real_port(opt(listener, "tcp-bound-port"), url)
     }
   }
-  send(sock, c(Sys.getpid(), url), mode = 2L, block = TRUE)
+  send(sock, url, mode = 2L, block = TRUE)
 
   ctx <- .context(sock)
   req <- recv_aio(ctx, mode = 8L, cv = cv)
@@ -109,9 +108,6 @@ dispatcher <- function(host, url = NULL, n = 0L, ...) {
             id <- as.character(-item)
             if (length(outq[[id]])) {
               outq[[id]][["msgid"]] && send(outq[[id]][["ctx"]], .connectionReset, mode = 1L, block = TRUE)
-              if (length(outq[[id]][["dmnid"]])) {
-                events <- c(events, outq[[id]][["dmnid"]])
-              }
               outq[[id]] <- NULL
             }
           }
@@ -132,16 +128,13 @@ dispatcher <- function(host, url = NULL, n = 0L, ...) {
               connections,
               awaiting,
               executing,
-              count - awaiting - executing,
-              events
+              count - awaiting - executing
             )
-            events <- integer()
           } else {
             found <- FALSE
             for (item in outq) {
               item[["msgid"]] == id && {
                 send(psock, 0L, mode = 1L, pipe = item[["pipe"]], block = TRUE)
-                `[[<-`(item, "msgid", -1L)
                 found <- TRUE
                 break
               }
@@ -169,24 +162,10 @@ dispatcher <- function(host, url = NULL, n = 0L, ...) {
         value <- .subset2(res, "value")
         id <- as.character(pipe_id(res))
         res <- recv_aio(psock, mode = 8L, cv = cv)
-        outq[[id]][["msgid"]] < 0 && {
-          `[[<-`(outq[[id]], "msgid", 0L)
-          cv_signal(cv)
-          next
-        }
         .read_marker(value) && {
-            send(outq[[id]][["ctx"]], value, mode = 2L, block = TRUE)
-            send(psock, 0L, mode = 2L, pipe = outq[[id]][["pipe"]], block = TRUE)
-            if (length(outq[[id]][["dmnid"]])) {
-              events <- c(events, outq[[id]][["dmnid"]])
-            }
-            outq[[id]] <- NULL
-          next
-        }
-        as.logical(value[1L]) || {
-          dmnid <- readBin(value, integer(), n = 2L)[2L]
-          events <- c(events, dmnid)
-          `[[<-`(outq[[id]], "dmnid", -dmnid)
+          send(outq[[id]][["ctx"]], value, mode = 2L, block = TRUE)
+          send(psock, 0L, mode = 2L, pipe = outq[[id]][["pipe"]], block = TRUE)
+          outq[[id]] <- NULL
           next
         }
         send(outq[[id]][["ctx"]], value, mode = 2L, block = TRUE)