+++ R/mirai-package.R
@@ -33,11 +33,11 @@
 #' `vignette("mirai", package = "mirai")`
 #'
 #' @importFrom nanonext .advance call_aio call_aio_ collect_aio collect_aio_
-#'   .context cv cv_signal cv_value dial ip_addr is_error_value .keep listen
-#'   .mark mclock monitor msleep nng_error opt opt<- parse_url pipe_id
+#'   .context cv cv_reset cv_signal cv_value dial ip_addr is_error_value .keep
+#'   listen .mark mclock monitor msleep nng_error opt opt<- parse_url pipe_id
 #'   pipe_notify random .read_header .read_marker read_monitor reap recv
 #'   recv_aio request send serial_config socket stat stop_aio stop_request
-#'   tls_config unresolved .unresolved until wait write_cert
+#'   tls_config unresolved .unresolved until wait wait_ write_cert
 #'
 "_PACKAGE"
 
@@ -45,7 +45,10 @@
 # tested implicitly
 
 .onLoad <- function(libname, pkgname) {
-  otel_tracing <<- requireNamespace("otel", quietly = TRUE) && otel::is_tracing_enabled()
+  otel_tracing <<- requireNamespace("otel", quietly = TRUE) && {
+    otel_tracer <<- otel::get_tracer(name = otel_tracer_name)
+    .subset2(otel_tracer, "is_enabled")()
+  }
   switch(
     Sys.info()[["sysname"]],
     Linux = {
@@ -77,7 +80,7 @@
   list(
     arglen = "`n` must equal the length of `args`, or either must be 1",
     cluster_inactive = "cluster is no longer active",
-    daemons_unset = "daemons must be set to use launchers",
+    daemons_unset = "daemons must be set to use this function",
     dot_required = "`.` must be an element of the character vector(s) supplied to `args`",
     function_required = "`.f` must be of type function, not %s",
     localhost = "SSH tunnelling requires daemons `url` hostname to be `127.0.0.1`",
@@ -89,10 +92,12 @@
     numeric_n = "`n` must be numeric, did you mean to provide `url`?",
     sync_daemons = "mirai: initial sync with daemon(s) [%d secs elapsed]",
     sync_dispatcher = "mirai: initial sync with dispatcher [%d secs elapsed]",
+    synchronous = "daemons cannot be launched for synchronous compute profiles",
     within_map = "cannot create local daemons from within mirai map"
   ),
   hash = TRUE
 )
 
 otel_tracing <- FALSE
+otel_tracer <- NULL
 otel_tracer_name <- "org.r-lib.mirai"